+++ dashboard/src/components/EventsContainer.js
@@ -14,18 +14,19 @@ import { MenuItem } from "material-ui/Menu";
 import { ListItemText, ListItemIcon } from "material-ui/List";
 import Checkbox from "material-ui/Checkbox";
 
-import EventsListItem from "./EventsListItem";
-import EventStatus from "./EventStatus";
-import ResolveEventMutation from "../mutations/ResolveEventMutation";
+import EventsListItem from "/components/EventsListItem";
+import EventStatus from "/components/EventStatus";
+import resolveEvent from "/mutations/resolveEvent";
+
 import TableList, {
   TableListHeader,
   TableListBody,
   TableListSelect,
   TableListEmptyState,
   TableListButton as Button,
-} from "./TableList";
+} from "/components/TableList";
 
-import Loader from "./Loader";
+import Loader from "/components/Loader";
 
 const styles = theme => {
   const toolbar = theme.mixins.toolbar;
@@ -187,7 +188,7 @@ class EventsContainer extends React.Component {
     );
 
     selectedKeys.forEach(key => {
-      ResolveEventMutation.commit(this.props.client, key).then(
+      resolveEvent(this.props.client, { id: key }).then(
         () => {
           this.setState(({ rowState }) =>
             Object.assign(rowState, { [key]: false }),