+++ dashboard/src/index.js
@@ -3,13 +3,13 @@ import ReactDOM from "react-dom";
 import { BrowserRouter } from "react-router-dom";
 import injectTapEventPlugin from "react-tap-event-plugin";
 
-import createClient from "./apollo/client";
+import createClient from "/apollo/client";
 
-import createStore from "./store";
-import reducer from "./reducer";
-import registerServiceWorker from "./registerServiceWorker";
+import createStore from "/store";
+import reducer from "/reducer";
+import registerServiceWorker from "/registerServiceWorker";
 
-import AppRoot from "./components/AppRoot";
+import AppRoot from "/components/AppRoot";
 
 // Fonts
 import "typeface-roboto"; // eslint-disable-line